fix: even stricter volume bounds - step matching
After checking the BINoculars source code again, I found that even stricter conditions for the reciprocal space boundaries are needed than I previously thought.
BINoculars not only (reasonably) expects volume_stop - volume_start to be an integer multiple of volume_step, but it also internally converts both limits to integer multiples of the step (i.e. volume_start/stop can only be integer*volume_step). While this is not strictly required for the read in HDF5 file, not following this convention possible leads to a mismatch of array sizes and a slight shifts of the reciprocal space volume.